refactor(common): Rename ttl_cache module to ttl
Since it doesn't contain the TtlCache anymore.
