+++ crates/matrix-sdk-indexeddb/src/state_store/mod.rs
@@ -37,7 +37,7 @@ use matrix_sdk_base::{
         SupportedVersionsResponse, ThreadSubscriptionStatus, WellKnownResponse,
         compare_thread_subscription_bump_stamps,
     },
-    ttl_cache::TtlValue,
+    ttl::TtlValue,
 };
 use matrix_sdk_store_encryption::{Error as EncryptionError, StoreCipher};
 use ruma::{
