+++ crates/matrix-sdk/src/authentication/oauth/tests.rs
@@ -2,7 +2,7 @@ use std::time::Duration;
 
 use anyhow::Context as _;
 use assert_matches::assert_matches;
-use matrix_sdk_base::{sleep::sleep, store::RoomLoadSettings, ttl_cache::TtlValue};
+use matrix_sdk_base::{sleep::sleep, store::RoomLoadSettings, ttl::TtlValue};
 use matrix_sdk_test::async_test;
 use oauth2::{ClientId, CsrfToken, PkceCodeChallenge, RedirectUrl, Scope};
 use ruma::{
