+++ crates/matrix-sdk/src/client/builder/mod.rs
@@ -28,7 +28,7 @@ use homeserver_config::*;
 use matrix_sdk_base::crypto::DecryptionSettings;
 #[cfg(feature = "e2e-encryption")]
 use matrix_sdk_base::crypto::{CollectStrategy, TrustRequirement};
-use matrix_sdk_base::{BaseClient, ThreadingSupport, store::StoreConfig, ttl_cache::TtlValue};
+use matrix_sdk_base::{BaseClient, ThreadingSupport, store::StoreConfig, ttl::TtlValue};
 use matrix_sdk_common::cross_process_lock::CrossProcessLockConfig;
 #[cfg(feature = "sqlite")]
 use matrix_sdk_sqlite::SqliteStoreConfig;
